feat(history): #WB2-1324, add history store and undo/redo hook
